requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Admin==1.6.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.1.1
email-validator==2.1.0
gunicorn==21.2.0
openai
APScheduler==3.10.4
